update to use current demo standards + .NET Core
