+++ .circleci/config.yml
@@ -1,39 +1,73 @@
 version: 2.1
 
 orbs:
-   win: circleci/windows@1.0.0
+   win: circleci/windows@2.4.0
 
 workflows:
   test:
     jobs:
-      - build-run-windows
+    - build-run-linux:
+        context: hello-world-demos
+    - build-run-windows:
+        context: hello-world-demos
+
+  test-daily:
+    triggers:
+    - schedule:
+        cron: "0 6 * * *"
+        filters:
+          branches:
+            only: master
+    jobs:
+    - build-run-linux:
+        context: hello-world-demos
+    - build-run-windows:
+        context: hello-world-demos
 
 # This CI build ensures that the demo both compiles and works correctly. For the runtime test,
 # we use an SDK key and flag key that are passed in via the CircleCI project configuration;
 # the flag is configured to return a true value.
 
 jobs:
+  build-run-linux:
+    docker:
+      - image: mcr.microsoft.com/dotnet/core/sdk:2.1-focal
+    steps:
+      - checkout
+      - run:
+          name: insert SDK key and flag key into demo code
+          command: |
+            sed -i.bak "s/SdkKey *= *\".*\"/SdkKey = \"${LD_HELLO_WORLD_SDK_KEY}\"/" ./HelloDotNet/Hello.cs
+            sed -i.bak "s/FeatureFlagKey *= *\".*\"/FeatureFlagKey = \"${LD_HELLO_WORLD_FLAG_KEY_WITH_TRUE_VALUE}\"/" ./HelloDotNet/Hello.cs
+      - run: dotnet build
+      - run:
+          name: run demo
+          command: |
+            dotnet run --project HelloDotNet | tee output.txt
+            grep "is True for this user" output.txt || (echo "Flag did not evaluate to expected true value" && exit 1)
+
   build-run-windows:
     executor:
-      name: win/vs2019
+      name: win/default
       shell: powershell.exe
     steps:
       - checkout
       - run:
           name: insert SDK key and flag key into demo code
           command: |
-            $content = Get-Content -Path .\HelloDotnet\Parameters.cs
-            $content = $content -Replace "SdkKey = `".*`"","SdkKey = `"$env:LD_DEMO_SDK_KEY`""
-            $content = $content -Replace "FeatureFlagKey = `".*`"","FeatureFlagKey = `"$env:LD_DEMO_FLAG_KEY`""
-            $content | Set-Content -Path .\HelloDotnet\Parameters.cs
+            $content = Get-Content -Path .\HelloDotNet\Hello.cs
+            $content = $content -Replace "SdkKey = `".*`"","SdkKey = `"$env:LD_HELLO_WORLD_SDK_KEY`""
+            $content = $content -Replace "FeatureFlagKey = `".*`"","FeatureFlagKey = `"$env:LD_HELLO_WORLD_FLAG_KEY_WITH_TRUE_VALUE`""
+            $content | Set-Content -Path .\HelloDotNet\Hello.cs
       - run:
           name: build demo
-          command: dotnet build HelloDotnet.sln
+          command: dotnet build HelloDotNet.sln
       - run:
           name: run demo
           command: |
-            Start-Transcript -Path .\output.txt -UseMinimalHeader
-            .\HelloDotNet\bin\Debug\HelloDotnet
-            Stop-Transcript
-            $output = Get-Content -Path .\output.txt
-            if (-Not $output.Contains("is True for this user")) { exit 1 }
+            .\HelloDotNet\bin\Debug\HelloDotnet | tee-object -variable output
+            if ( "$output" -notlike "*is True for this user*" )
+            {
+              write-output "Flag did not evaluate to expected true value"
+              exit 1
+            }
