--- HelloDotNet/Parameters.cs
@@ -1,12 +1,0 @@
-
-namespace HelloDotNet
-{
-    public class Parameters
-    {
-        // Set SdkKey to your LaunchDarkly SDK key before compiling
-        public const string SdkKey = "";
-
-        // Set FEATURE_FLAG_KEY to the feature flag key you want to evaluate
-        public const string FeatureFlagKey = "";
-    }
-}
