--- HelloDotNet/packages.config
@@ -1,51 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Common.Logging" version="3.4.1" targetFramework="net451" />
-  <package id="Common.Logging.Core" version="3.4.1" targetFramework="net451" />
-  <package id="LaunchDarkly.Cache" version="1.0.2" targetFramework="net451" />
-  <package id="LaunchDarkly.CommonSdk.StrongName" version="2.11.0" targetFramework="net451" />
-  <package id="LaunchDarkly.EventSource" version="3.3.0" targetFramework="net451" />
-  <package id="LaunchDarkly.ServerSdk" version="5.13.0" targetFramework="net451" />
-  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net452" />
-  <package id="Microsoft.Bcl" version="1.1.10" targetFramework="net452" />
-  <package id="Microsoft.Bcl.Build" version="1.0.21" targetFramework="net452" />
-  <package id="Microsoft.CSharp" version="4.0.1" targetFramework="net451" />
-  <package id="Microsoft.Net.Http" version="2.2.29" targetFramework="net452" />
-  <package id="Microsoft.NETCore.Platforms" version="1.1.0" targetFramework="net451" />
-  <package id="NETStandard.Library" version="1.6.1" targetFramework="net451" />
-  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net45" />
-  <package id="NuGet.CommandLine" version="3.5.0" targetFramework="net451" developmentDependency="true" />
-  <package id="System.Collections" version="4.3.0" targetFramework="net451" />
-  <package id="System.Collections.Concurrent" version="4.3.0" targetFramework="net451" />
-  <package id="System.Collections.Immutable" version="1.2.0" targetFramework="net451" />
-  <package id="System.ComponentModel" version="4.3.0" targetFramework="net451" />
-  <package id="System.Diagnostics.Debug" version="4.3.0" targetFramework="net451" />
-  <package id="System.Diagnostics.Tools" version="4.3.0" targetFramework="net451" />
-  <package id="System.Diagnostics.Tracing" version="4.3.0" targetFramework="net451" />
-  <package id="System.Globalization" version="4.3.0" targetFramework="net451" />
-  <package id="System.IO" version="4.3.0" targetFramework="net451" />
-  <package id="System.IO.Compression" version="4.3.0" targetFramework="net451" />
-  <package id="System.Linq" version="4.3.0" targetFramework="net451" />
-  <package id="System.Linq.Expressions" version="4.3.0" targetFramework="net451" />
-  <package id="System.Net.Http" version="4.3.2" targetFramework="net451" />
-  <package id="System.Net.Primitives" version="4.3.0" targetFramework="net451" />
-  <package id="System.ObjectModel" version="4.3.0" targetFramework="net451" />
-  <package id="System.Reflection" version="4.3.0" targetFramework="net451" />
-  <package id="System.Reflection.Extensions" version="4.3.0" targetFramework="net451" />
-  <package id="System.Reflection.Primitives" version="4.3.0" targetFramework="net451" />
-  <package id="System.Resources.ResourceManager" version="4.3.0" targetFramework="net451" />
-  <package id="System.Runtime" version="4.3.0" targetFramework="net451" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.3.0" targetFramework="net451" />
-  <package id="System.Runtime.Extensions" version="4.3.0" targetFramework="net451" />
-  <package id="System.Runtime.InteropServices" version="4.3.0" targetFramework="net451" />
-  <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.3.0" targetFramework="net451" />
-  <package id="System.Runtime.Numerics" version="4.3.0" targetFramework="net451" />
-  <package id="System.Text.Encoding" version="4.3.0" targetFramework="net451" />
-  <package id="System.Text.Encoding.Extensions" version="4.3.0" targetFramework="net451" />
-  <package id="System.Text.RegularExpressions" version="4.3.0" targetFramework="net451" />
-  <package id="System.Threading" version="4.3.0" targetFramework="net451" />
-  <package id="System.Threading.Tasks" version="4.3.0" targetFramework="net451" />
-  <package id="System.Threading.Timer" version="4.3.0" targetFramework="net451" />
-  <package id="System.Xml.ReaderWriter" version="4.3.0" targetFramework="net451" />
-  <package id="System.Xml.XDocument" version="4.3.0" targetFramework="net451" />
-</packages>
